+++ Makefile
@@ -1,57 +1,87 @@
-makebase = if [ !  -d ~/.tapp ]; \
+PYTHON_VERSION=2.7  # not used...
+SHELL = /bin/bash  # for pushd, but would probably be better to avoid this
+PREREQDIR=.
+
+makelog = if [ !  -d /var/log/trademanager/ ]; \
 	then \
-		mkdir ~/.tapp; \
+		sudo mkdir /var/log/trademanager/; \
 	fi
 
-makedirs = if [ !  -d ~/.tapp/trademanager ]; \
+makerun = if [ !  -d /var/run/trademanager/ ]; \
 	then \
-		mkdir ~/.tapp/trademanager; \
-		mkdir ~/.tapp/helper; \
-		cp cfg.ini ~/.tapp/trademanager; \
-		cp cfg.ini ~/.tapp/helper; \
+		sudo mkdir /var/run/trademanager/; \
 	fi
 
-installprereqs = if [ !  -d $(1)ledger ]; \
-    git clone git://github.com/ledger/ledger.git \
-    pushd $(1)ledger \
-    ./acprep --python update \
-    popd \
-    export PYTHONPATH=$PYTHONPATH:$(readlink -f $(1)ledger) \
-    if [ !  -d $(1)libsecp256k1 ]; \
-    git clone git://github.com/bitcoin/secp256k1.git libsecp256k1 \
-    pushd $(1)libsecp256k1 \
-    git checkout d7eb1ae96dfe9d497a26b3e7ff8b6f58e61e400a \
-    ./autogen.sh \
-    ./configure --enable-module-recovery --enable-module-ecdh --enable-module-schnorr \
-    make \
-    make install \
-    popd \
-    INCLUDE_DIR=$(readlink -f $(1)libsecp256k1/include) \
-    LIB_DIR=$(readlink -f $(1)libsecp256k1/.libs) \
-    LD_LIBRARY_PATH=$(readlink -f $(1)libsecp256k1/.libs)
+installledger = if [ !  -d $(1)/ledger ]; \
+	then \
+		git clone git://github.com/ledger/ledger.git $(1)/ledger; \
+		pushd $(1)/ledger; \
+		if [ $(VIRTUAL_ENV) ]; \
+			then \
+				./acprep --prefix=$(VIRTUAL_ENV) --python update; \
+			else \
+				./acprep  --python update; \
+		fi; \
+		make; \
+		make install; \
+		popd; \
+	fi
+
+installsecp256k1 = if [ !  -d $(1)/libsecp256k1 ]; \
+	then \
+		git clone git://github.com/bitcoin/secp256k1.git $(1)/libsecp256k1; \
+		pushd $(1)/libsecp256k1; \
+		git checkout d7eb1ae96dfe9d497a26b3e7ff8b6f58e61e400a; \
+		./autogen.sh; \
+		if [ $(VIRTUAL_ENV) ]; \
+			then \
+				./configure --enable-module-recovery --enable-module-ecdh --enable-module-schnorr --prefix=$(VIRTUAL_ENV); \
+			else \
+				./configure --enable-module-recovery --enable-module-ecdh --enable-module-schnorr; \
+		fi; \
+		make; \
+		make install; \
+		popd; \
+	fi
+
+installlocalgit= if [ -d $(1) ]; \
+	then \
+		pushd $(1); \
+		python setup.py install; \
+		popd; \
+	else \
+		git clone $(2) $(1); \
+		pushd $(1); \
+		python setup.py install; \
+		popd; \
+	fi
+
+makeclean = rm -rf .cache build dist *.egg-info test/__pycache__ \
+	rm -rf test/*.pyc *.egg *~ *pyc test/*~ .eggs \
+	rm -f .coverage*
 
 build:
 	python setup.py build
 
+hackpendingPRs:
+	$(call installlocalgit, $(shell readlink -f ../../tapp/tappmq/ https://github.com/GitGuild/tappmq.git))
+	$(call installlocalgit, $(shell readlink -f ../../tapp/sqlalchemy-models/ https://github.com/GitGuild/sqlalchemy-models.git))
+
+dependencies:
+	$(call installledger, $(PREREQDIR))
+	$(call installsecp256k1, $(PREREQDIR))
+
 install:
-	$(call makebase, "")
-	$(call makedirs, "")
-    #if [ $(target) ] \
-    #then \
-    #    $(call installprereqs, $(target)) \
-	#else \
-	#    $(call installprereqs, ./) \
-    #fi
+	$(call makelog, "")
+	$(call makerun, "")
 	python setup.py -v install
 
 clean:
-	rm -rf .cache build dist *.egg-info test/__pycache__
-	rm -rf test/*.pyc *.egg *~ *pyc test/*~ .eggs
-	rm -f .coverage*
+	$(call makeclean, "")
 
 purge:
-	rm -rf .cache build dist *.egg-info test/__pycache__
-	rm -rf test/*.pyc *.egg *~ *pyc test/*~ .eggs
-	rm -f .coverage*
-	rm -rf ~/.tapp/trademanager
-	rm -rf ~/.tapp/helper
+	$(call makeclean, "")
+	sudo rm -rf /var/run/trademanager
+	sudo rm -rf /var/log/trademanager
+	rm -rf $(PREREQDIR)/ledger
+	rm -rf $(PREREQDIR)/libsecp256k1
