+++ example_cfg.ini
@@ -1,0 +1,27 @@
+[db]
+SA_ENGINE_URI: postgresql://postgres@trading
+
+[log]
+LOGFILE:
+LOGLEVEL: INFO
+
+[helper]
+KEY: pubkey
+SECRET: secret
+active_markets: ["BTC_USD", "DASH_BTC"]
+userpubkey: 1asdfasdfasdf
+
+[poloniex]
+KEY: key
+SECRET: secret
+active_markets: ["BTC_USD", "DASH_BTC", "DASH_USD", "ETH_BTC", "ETH_USD", "LTC_BTC", "LTC_USD", "LSK_BTC", "FCT_BTC", "LBC_BTC", "ETC_BTC", "ETC_USD", "OMNI_BTC"]
+
+[bitfinex]
+KEY: key
+SECRET: secret
+active_markets: ["BTC_USD", "ETH_USD", "ETH_BTC", "LTC_BTC", "LTC_USD", "BFX_USD", "BFX_BTC", "ETC_BTC", "ETC_USD"]
+
+[kraken]
+KEY: key
+SECRET: secret
+active_markets: ["BTC_USD", "ETH_USD", "ETH_BTC", "ETC_BTC", "ETC_USD"]
