+++ supervisord.conf
@@ -1,45 +1,53 @@
 [unix_http_server]
-file = /tmp/supervisor_helper.sock
+file = /tmp/supervisor_tm.sock
 
 [supervisorctl]
-serverurl = unix:///tmp/supervisor_helper.sock
+serverurl = unix:///tmp/supervisor_tm.sock
 
 [rpcinterface:supervisor]
 supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
 
 [supervisord]
-logfile = /home/ira/.tapp/supervisord.log
-childlogdir  = /home/ira/.tapp/
-pidfile = /home/ira/.tapp/supervisord.pid
+logfile = /var/log/trademanager/supervisord.log
+childlogdir  = /var/log/trademanager/
+pidfile = /var/run/trademanager/supervisord.pid
 loglevel = debug
-
-[program:helper]
-command=python trade_manager/helper.py
-autostart=false
 
 [group:bitfinex]
 programs=bitfinexm,bitfinexl
-autostart=false
 
 [program:bitfinexm]
 command=bitfinexm
+autorestart=true
+stopasgroup=true
+autostart=false
 
 [program:bitfinexl]
 command=bitfinexl
+autorestart=true
+stopasgroup=true
+autostart=false
 
 [program:kraken]
 command=krakenm
+autorestart=true
+stopasgroup=true
 autostart=false
 
 [group:poloniex]
 programs=poloniexm,poloniexl
-autostart=false
 
 [program:poloniexm]
 command=poloniexm
+autorestart=true
+stopasgroup=true
+autostart=false
 
 [program:poloniexl]
 command=poloniexl
+autorestart=true
+stopasgroup=true
+autostart=false
 
 [eventlistener:trade_listener]
 command=tapplistener
