+++ test_supervisord.conf
@@ -1,0 +1,23 @@
+[unix_http_server]
+file = /tmp/supervisor_helper.sock
+
+[supervisorctl]
+serverurl = unix:///tmp/supervisor_helper.sock
+
+[rpcinterface:supervisor]
+supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
+
+[supervisord]
+logfile = /tmp/test_supervisord.log
+childlogdir  = /tmp
+pidfile = /tmp/test_supervisord.pid
+loglevel = debug
+
+[program:helper]
+command=python test/helper.py
+autostart=false
+
+[eventlistener:trade_listener]
+command=tapplistener
+events=PROCESS_STATE
+buffer_size=100
