+++ trade_manager/plugin.py
@@ -1,14 +1,11 @@
 import collections
+import datetime
 import json
 import time
+from ledger import Amount
 from ledger import commodities, Balance
 
-import datetime
-
-from alchemyjsonschema.dictify import jsonify
-from ledger import Amount
-
-from sqlalchemy_models import sa, Base, create_session_engine, jsonify2
+from sqlalchemy_models import sa, Base, create_session_engine
 from sqlalchemy_models.util import filter_query_by_attr, multiply_tickers
 from tapp_config import get_config, setup_redis
 from tappmq.tappmq import publish, MQHandlerBase, get_running_workers
